# [#714] Improve map get performance