--- a/scripts/profile-function
+++ b/scripts/profile-function
@@ -59,7 +59,7 @@ fi
 ################################################################################
 
 # Fetch dev deps and compile
-rebar3 as dev do upgrade, compile
+rebar3 as dev do upgrade, clojerl compile
 
 # Generate args
 ARGS=$(echo "$@" | tr ' ' ,)
@@ -71,7 +71,7 @@ SVG_FILE="$TRACE_FILE_SORTED.svg"
 
 echo "Running eflame for $MODULE:$FUNCTION($ARGS)"
 
-PID=$(erl -sname eflame-expr -pa _build/*/lib/*/ebin -pa ebin -s clojerl -noshell \
+PID=$(erl -sname eflame-expr -pa _build/dev/lib/*/ebin -pa ebin -s clojerl -noshell \
           -eval "erlang:apply('$MODULE', '$FUNCTION', [$ARGS]), \
                  eflame:apply(normal_with_children, \"$TRACE_FILE\", '$MODULE', '$FUNCTION', [$ARGS]), \
                  io:format(\"~p~n\", [self()]), \