Fixes #714 

### (Micro)Benchmark code

```
(simple-benchmark [coll {:foo 1 :bar 2}] (get coll :foo) 1000000)
(simple-benchmark [coll {'foo 1 'bar 2}] (get coll 'foo) 1000000)
(simple-benchmark [coll {:foo 1 :bar 2}] (:foo coll) 1000000)
(simple-benchmark [coll {'foo 1 'bar 2}] ('foo coll) 1000000)
(let [kw  :foo
      sym 'foo]
  (simple-benchmark [coll {:foo 1 :bar 2}] (kw coll) 1000000)
  (simple-benchmark [coll {'foo 1 'bar 2}] (sym coll) 1000000))
```

### Before

```
[coll {:foo 1, :bar 2}], (get coll :foo), 1000000 runs, 148 msecs
[coll {(quote foo) 1, (quote bar) 2}], (get coll (quote foo)), 1000000 runs, 202 msecs
[coll {:foo 1, :bar 2}], (:foo coll), 1000000 runs, 172 msecs
[coll {(quote foo) 1, (quote bar) 2}], ((quote foo) coll), 1000000 runs, 213 msecs
[coll {:foo 1, :bar 2}], (kw coll), 1000000 runs, 175 msecs
[coll {(quote foo) 1, (quote bar) 2}], (sym coll), 1000000 runs, 217 msecs
```

### After

```
[coll {:foo 1, :bar 2}], (get coll :foo), 1000000 runs, 112 msecs
[coll {(quote foo) 1, (quote bar) 2}], (get coll (quote foo)), 1000000 runs, 171 msecs
[coll {:foo 1, :bar 2}], (:foo coll), 1000000 runs, 142 msecs
[coll {(quote foo) 1, (quote bar) 2}], ((quote foo) coll), 1000000 runs, 192 msecs
[coll {:foo 1, :bar 2}], (kw coll), 1000000 runs, 140 msecs
[coll {(quote foo) 1, (quote bar) 2}], (sym coll), 1000000 runs, 191 msecs
```